+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Version 4.1.0-dev
 - added assertions assertGreaterThan and assertLessThan
 - moved interfaces ICodeCoverageEngine and ICodeCoverageFormatter to namespace MyTester\CodeCoverage
-- moved interfaces IAnnotationsReaderEngine to namespace MyTester\Annotations
+- moved interface IAnnotationsReaderEngine to namespace MyTester\Annotations
 - added option to mark a test as incomplete
 
 Version 4.0.0